refactor(shared): extract formatRelativeTime into shared utility
Replaces 5 duplicate implementations of German relative time formatting
across apps/web, apps/mobile, packages/docs, and packages/chat.

New shared utility in packages/shared/src/utils/formatRelativeTime.ts
accepts string | number | Date input and supports configurable maxDays
and date fallback format options.
