+++ packages/docs/src/components/chat/ChatMessage.tsx
@@ -1,5 +1,6 @@
 import { memo } from 'react';
 import { Avatar, AvatarFallback } from '@gruenerator/ui';
+import { formatRelativeTime } from '@gruenerator/shared/utils';
 import type { ChatMessage as ChatMessageType } from '../../hooks/useDocumentChat';
 
 function getInitials(name: string): string {
@@ -8,24 +9,6 @@ function getInitials(name: string): string {
     return (parts[0][0] + parts[parts.length - 1][0]).toUpperCase();
   }
   return name.slice(0, 2).toUpperCase();
-}
-
-function formatRelativeTime(timestamp: number): string {
-  const diff = Date.now() - timestamp;
-  const seconds = Math.floor(diff / 1000);
-  const minutes = Math.floor(seconds / 60);
-  const hours = Math.floor(minutes / 60);
-
-  if (seconds < 60) return 'Gerade eben';
-  if (minutes < 60) return `vor ${minutes} Min.`;
-  if (hours < 24) return `vor ${hours} Std.`;
-
-  return new Date(timestamp).toLocaleDateString('de-DE', {
-    day: '2-digit',
-    month: '2-digit',
-    hour: '2-digit',
-    minute: '2-digit',
-  });
 }
 
 interface ChatMessageProps {
